requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.0
requests==2.31.0
gunicorn==21.2.0
